+++ src/main/java/com/inditex/logandotest/api/model/Stock.java
@@ -9,13 +9,13 @@ import lombok.Data;
 public class Stock {
 
     @Id
-    private int sizeId;
+    private Long sizeId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "size_id", referencedColumnName = "id", nullable = false)
     private Size size;
 
     @Column(nullable = false)
-    private int quantity;
+    private Integer quantity;
 
 }
